requirement.txt
torch
torchvision
accelerate
diffusers
transformers
safetensors
Pillow
streamlit
